attempt to avoid js library conflicts.
a number of other javascript libraries (e.g. prototype and mootools)
also make use of $ as a shortcut alias. calling noConflict() after
jQuery is (hopefully) loaded should eliminate most of these, but
won't address the case where we have a page that uses a very old
version of jQuery. this can be worked around as well but requires
that we produce our own jQuery build using a reasonably unique
global identifier (i.e. not jQuery or $).